+++ src/test/java/cd/go/contrib/elasticagents/dockerswarm/elasticagent/executors/ProfileValidateRequestExecutorTest.java
@@ -16,25 +16,58 @@
 
 package cd.go.contrib.elasticagents.dockerswarm.elasticagent.executors;
 
+import cd.go.contrib.elasticagents.dockerswarm.elasticagent.PluginRequest;
+import cd.go.contrib.elasticagents.dockerswarm.elasticagent.PluginSettings;
 import cd.go.contrib.elasticagents.dockerswarm.elasticagent.requests.ProfileValidateRequest;
+import com.thoughtworks.go.plugin.api.response.GoPluginApiResponse;
 import org.junit.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
 import java.util.Collections;
+import java.util.HashMap;
+
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class ProfileValidateRequestExecutorTest {
     @Test
     public void shouldBarfWhenUnknownKeysArePassed() throws Exception {
-        ProfileValidateRequestExecutor executor = new ProfileValidateRequestExecutor(new ProfileValidateRequest(Collections.singletonMap("foo", "bar")));
+        ProfileValidateRequestExecutor executor = new ProfileValidateRequestExecutor(new ProfileValidateRequest(Collections.singletonMap("foo", "bar")), null);
         String json = executor.execute().responseBody();
         JSONAssert.assertEquals("[{\"message\":\"Image must not be blank.\",\"key\":\"Image\"},{\"key\":\"foo\",\"message\":\"Is an unknown property\"}]", json, JSONCompareMode.NON_EXTENSIBLE);
     }
 
     @Test
     public void shouldValidateMandatoryKeys() throws Exception {
-        ProfileValidateRequestExecutor executor = new ProfileValidateRequestExecutor(new ProfileValidateRequest(Collections.<String, String>emptyMap()));
+        ProfileValidateRequestExecutor executor = new ProfileValidateRequestExecutor(new ProfileValidateRequest(Collections.<String, String>emptyMap()), null);
         String json = executor.execute().responseBody();
         JSONAssert.assertEquals("[{\"message\":\"Image must not be blank.\",\"key\":\"Image\"}]", json, JSONCompareMode.NON_EXTENSIBLE);
+    }
+
+    @Test
+    public void shouldValidateDockerSecrets() throws Exception {
+        final HashMap<String, String> properties = new HashMap<>();
+        properties.put("Image", "alpine");
+        properties.put("Secrets", "Foo");
+        final PluginRequest pluginRequest = mock(PluginRequest.class);
+        final PluginSettings pluginSettings = mock(PluginSettings.class);
+
+        when(pluginRequest.getPluginSettings()).thenReturn(pluginSettings);
+
+        GoPluginApiResponse response = new ProfileValidateRequestExecutor(new ProfileValidateRequest(properties), pluginRequest).execute();
+
+        assertThat(response.responseCode(), is(200));
+
+        final String expectedJson = "[\n" +
+                "  {\n" +
+                "    \"key\": \"Secrets\",\n" +
+                "    \"message\": \"Invalid secret specification `Foo`. Property `src` is required.\"\n" +
+                "  }\n" +
+                "]";
+
+        JSONAssert.assertEquals(expectedJson, response.responseBody(), true);
     }
 }